\name{dmGsea}
\alias{dmGsea}
\title{
	Efficient Gene Set Enrichment Analysis for DNA Methylation Data
	
}
\description{
The R package dmGsea is specifically designed for DNA methylation data, provides 
functions to perform gene set enrichment analysis while addressing probe 
dependency and probe number bias. The package supports Illumina 450K, EPIC, 
and mouse methylation arrays and can be extended to other omics data with 
user-provided probe-to-gene mapping annotations. It has four main functions
to perform gene set enrichment analysis. gsGene: GSEA based on aggregates 
association signals at gene level;gsPG:  GSEA using summary statistics 
for independent probe groups based on gene annotation; gsProbe: GSEA using 
probe level p-values; gsRank: Fast ranking based GSEA based on gene level 
statistics
}


\author{
Zongli Xu
}


